# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AgentoGo API",
        "description": "API REST multi-tenant con autenticación, auditoría y gestión de negocios",
        "contact": {
            "name": "AgentoGo Support",
            "email": "support@agentogo.com"
        },
        "version": "0.3.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/businesses/{business_id}/reservations": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Listar reservas del negocio",
                "operationId": "4494a89f4d98d7afd4f97faa52755ab1",
                "responses": {
                    "200": {
                        "description": "Lista de reservas con paginación"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Crear una reserva",
                "operationId": "99e31d4328db123c2355bba0a2748239",
                "responses": {
                    "201": {
                        "description": "Reserva creada"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "start_time": [
                                            "La fecha/hora de inicio es requerida."
                                        ],
                                        "employee_id": [
                                            "Empleado no encontrado o no disponible."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/businesses/{business_id}/reservations/{reservation_id}": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Obtener una reserva",
                "operationId": "94b9c9d1e94a56ab828cc3e31aa95afa",
                "responses": {
                    "200": {
                        "description": "Detalle de la reserva"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Actualizar una reserva",
                "operationId": "daaedac3cb021277edc496b231cbee44",
                "responses": {
                    "200": {
                        "description": "Reserva actualizada"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "start_time": [
                                            "La fecha/hora de inicio es requerida."
                                        ],
                                        "employee_id": [
                                            "Empleado no encontrado o no disponible."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Eliminar una reserva",
                "operationId": "a1b8beebf99146188b960f93cde6358d",
                "responses": {
                    "204": {
                        "description": "Reserva eliminada"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Error403": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 403
                    },
                    "message": {
                        "type": "string",
                        "example": "No autorizado"
                    }
                },
                "type": "object",
                "example": {
                    "code": 403,
                    "message": "No tiene permisos para realizar esta acción."
                }
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Mapa de errores por campo",
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "message": "The given data was invalid.",
                    "errors": {
                        "name": [
                            "El campo nombre es requerido."
                        ],
                        "email": [
                            "El email debe ser una dirección válida."
                        ]
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación Sanctum",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}